# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Vulpi API",
    "version": "1.0.0",
    "description": "API pro integraci s fakturačním systémem Vulpi."
  },
  "servers": [
    {
      "url": "https://app.vulpi.cz/api/v1",
      "description": "Produkční server"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
    "schemas": {
      "Invoice": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "number": {
            "type": "string",
            "example": "2023001"
          },
          "status": {
            "type": "string",
            "enum": ["DRAFT", "ISSUED", "PAID", "OVERDUE", "CANCELLED"]
          },
          "totalAmount": {
            "type": "number",
            "format": "float"
          },
          "currency": {
            "type": "string",
            "example": "CZK"
          },
          "variableSymbol": {
            "type": "string"
          },
          "issuedAt": {
            "type": "string",
            "format": "date-time"
          },
          "dueDate": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "paths": {
    "/invoices": {
      "get": {
        "summary": "Seznam faktur",
        "description": "Vrátí seznam všech faktur organizace.",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": ["DRAFT", "ISSUED", "PAID", "OVERDUE", "CANCELLED"]
            },
            "description": "Filtr podle stavu faktury"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            },
            "description": "Počet záznamů na stránku"
          }
        ],
        "responses": {
          "200": {
            "description": "Úspěšná odpověď",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Invoice"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Neautorizovaný přístup"
          }
        }
      },
      "post": {
        "summary": "Vytvořit fakturu",
        "description": "Vytvoří novou fakturu.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["clientId", "items"],
                "properties": {
                  "clientId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "currency": {
                    "type": "string",
                    "default": "CZK"
                  },
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": ["description", "amount", "quantity"],
                      "properties": {
                        "description": { "type": "string" },
                        "amount": { "type": "number" },
                        "quantity": { "type": "number", "default": 1 },
                        "vatRate": { "type": "number", "default": 0 }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Faktura vytvořena",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Invoice"
                }
              }
            }
          }
        }
      }
    },
    "/invoices/{id}": {
      "get": {
        "summary": "Detail faktury",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detail faktury",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Invoice"
                }
              }
            }
          },
          "404": {
            "description": "Faktura nenalezena"
          }
        }
      }
    }
  }
}